+++ R/smooth3.R
@@ -1,0 +1,9 @@
+
+#' @export
+smooth3 <- function(data, filt = "b", sz = 3) {
+    padsize <- (sz - 1) / 2
+    ones(sz) / prod(sz)
+}
+
+
+