+++ R/zeros.R
@@ -7,7 +7,8 @@ m.zeros <- function(x, ...) {
     .CallOctave("zeros", x, ...)
 }
 
-#' Fills an array with zeros
+#' Fills an array with zeros as in Matlab
+#'
 #' @param x 1-D dimension
 #' @param ... additional parameters
 #' @export
@@ -15,10 +16,10 @@ zeros <- function(x, ...) {
     stopifnot(is.numeric(x))
     args <- list(...)
     if (length(args)) {
-        stopifnot(is.numeric(args[[1]]), is.numeric(args[[2]]))
+        stopifnot(sapply(args,  is.numeric))
         array(0, dim=c(x, ...))
     } else {
-        array(0, dim=x)
+        array(0, dim=c(x,x))
     }
 }