+++ NEWS.md
@@ -1,3 +1,11 @@
+## 20170729 0.0.1.9000
+* implemented functions ones, zeros, m.convn, m.spdiags (sparse diagonals)
+* unit test working for ones, zeros
+* TPFA function for MRST working in R
+* smooth3.m function under inst/matlab
+* class R6 Grid working
+* class R6 Var3D working
+
 ## 20170724 0.0.1
 * couple of notebooks
 * functions m.zeros, m.ones, m.spdiags